refactor(payment): Phase 3 IPaymentStore + IPaymentGateway to Domain/Abstractions/
- Move IPaymentStore + IPaymentGateway (with companion PaymentGatewayResult)
  into Payment.Service.Domain.Abstractions; PaymentContext still implements
  IPaymentStore. InMemoryPaymentGateway stays in Infrastructure/Gateways/.
- Update consumer usings across endpoints, event handlers, Program.cs,
  PaymentContext, and EntityFrameworkExtensions.
