+++ payment-microservice/Payment.Service/IntegrationEvents/EventHandlers/AuthorizePaymentCommandHandler.cs
@@ -5,8 +5,7 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Payment.Service.Contracts.Integration;
 using Payment.Service.Domain;
-using Payment.Service.Infrastructure.Data;
-using Payment.Service.Infrastructure.Gateways;
+using Payment.Service.Domain.Abstractions;
 using Payment.Service.Observability;
 
 namespace Payment.Service.IntegrationEvents.EventHandlers;
