+++ payment-microservice/Payment.Service/Program.cs
@@ -8,8 +8,8 @@ using ECommerce.Shared.Observability;
 using ECommerce.Shared.OpenApi;
 using ECommerce.Shared.Qa;
 using Payment.Service.Contracts.Integration;
+using Payment.Service.Domain.Abstractions;
 using Payment.Service.Endpoints;
-using Payment.Service.Infrastructure.Data;
 using Payment.Service.Infrastructure.Data.EntityFramework;
 using Payment.Service.Infrastructure.Gateways;
 using Payment.Service.IntegrationEvents.EventHandlers;
